[fix] 特定のノートのデコードに失敗する問題を修正しました。

+++ app/src/main/java/jp/panta/misskeyandroidclient/api/drive/FilePropertyDTO.kt
@@ -32,8 +32,8 @@ data class FilePropertyDTO(
 
     @kotlinx.serialization.Serializable
     data class Properties(
-        val width: Int? = null,
-        val height: Int? = null
+        val width: Float? = null,
+        val height: Float? = null
     ) : Serializable
 
     private fun getThumbnailUrl(instanceBaseUrl: String): String{
